Askademia/backend/requirements.txt
uagents
uagents-ai-engine
pymongo
requests
python-dotenv